multi select list for static data created
