+++ src/app/pages/dashboard/dashboard.html
@@ -159,3 +159,4 @@
 </div>
   </div>
 </div>
+
