+++ src/index.html
@@ -18,6 +18,8 @@
   <base href="<%= htmlWebpackPlugin.options.metadata.baseUrl %>">
 <!--   <link rel="stylesheet" href="app/theme/record-picker.css">
  -->
+ <link rel="stylesheet" href="css/record-picker.css">
+
 </head>
 
 <body>
